backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-generativeai==0.8.6
sqlalchemy==2.0.30
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-multipart==0.0.9
bcrypt>=4.1.0
